requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
joblib==1.4.2
pymongo==4.11.1
faker==37.1.0